hpp refactor
