+++ firmware/shared/srcpp/app/app_stackMonitor.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <array>
 #include <cstdint>
-#include "io_stackMonitor.h"
+#include "io_stackMonitor.hpp"
 
 namespace app
 {
