+++ firmware/shared/srcpp/app/app_timer.cpp
@@ -1,5 +1,5 @@
-#include "app_timer.h"
-#include "io_time.h"
+#include "app_timer.hpp"
+#include "io_time.hpp"
 
 namespace app
 {
