+++ firmware/shared/srcpp/app/app_utils.hpp
