+++ firmware/shared/srcpp/hw/hw_adc.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "app_utils.h"
+#include "app_utils.hpp"
 #include <array>
 
 #ifdef TARGET_EMBEDDED
