+++ firmware/shared/srcpp/hw/hw_can.hpp
@@ -2,7 +2,7 @@
 
 #include <cstdint>
 #ifdef TARGET_EMBEDDED
-#include "hw_hal.h"
+#include "hw_hal.hpp"
 #endif
 
 namespace hw::can
