+++ firmware/shared/srcpp/hw/hw_error.hpp
