+++ firmware/shared/srcpp/hw/hw_hal.hpp
