+++ firmware/shared/srcpp/hw/hw_hardFaultHandler.hpp
