+++ firmware/shared/srcpp/io/io_heartbeatNode.cpp
@@ -1,1 +1,1 @@
-#include "io_heartbeatNode.h"
+#include "io_heartbeatNode.hpp"
