+++ firmware/shared/srcpp/io/io_jsoncan.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "io_canMsgQueue.h"
+#include "io_canMsgQueue.hpp"
 
 extern "C"
 {
