+++ firmware/shared/srcpp/io/io_led.cpp
@@ -1,4 +1,4 @@
-#include "io_led.h"
+#include "io_led.hpp"
 
 void io::binary_led::set(bool state) const
 {
