+++ firmware/shared/srcpp/io/io_led.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "hw_gpio.h"
+#include "hw_gpio.hpp"
 
 namespace io
 {
