+++ firmware/shared/srcpp/io/io_log.cpp
@@ -1,4 +1,4 @@
-#include "io_log.h"
+#include "io_log.hpp"
 extern "C"
 {
 #include "app_canAlerts.h"
