+++ firmware/shared/srcpp/io/io_log.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "app_utils.h"
+#include "app_utils.hpp"
 
 /*
 This logging module users SEGGER Real-Time Transfer (RTT) for printf-style debugging.
