+++ firmware/shared/srcpp/io/io_stackMonitor.cpp
@@ -1,4 +1,4 @@
-#include "io_stackMonitor.h"
+#include "io_stackMonitor.hpp"
 
 namespace io
 {
