+++ firmware/shared/srcpp/io/io_switch.cpp
@@ -1,4 +1,4 @@
-#include "io_switch.h"
+#include "io_switch.hpp"
 
 namespace io
 {
