+++ firmware/shared/srcpp/io/io_time.cpp
@@ -1,4 +1,4 @@
-#include "io_time.h"
+#include "io_time.hpp"
 #include <cmsis_os.h>
 
 namespace io::time
